requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-adk==0.3.0